Conflicts:
	src/application/enseignant.php
	src/application/etudiant.php
	src/ressources/classes/Etudiant.php
	src/ressources/classes/Voeu.php

+++ src/application/ajax/deleteTache.php
@@ -14,19 +14,25 @@ if (isset($_POST))
 {
 	//création des DAO
 	$tachesDAO = new TachesDAO ( MaBD::getInstance () );
-	$realisesDAO = new RealisesDAO ( MaBD::getInstance () );
 
 	//récupération des données
 	$num = $_POST['no_tache'];
-	$etudiants = $_POST['etudiants'];
 
 	//supression de la tache
 	$tacheDel= $tachesDAO->getOne($num);
-	$tachesDAO->delete($tacheDel);
-	echo json_encode(true);
+	if ($tacheDel!=null) 
+	{
+		$tachesDAO->delete($tacheDel);
+		echo json_encode(true);
+	}
+	else
+	{
+		echo json_encode('Désolé une erreur est survenue si celle-ci persiste veuillez la signaler');
+	}
+
 }
 else
 {
-	echo json_encode('Désolée une erreur est survenu si celle-ci persiste veuillez la signaler');
+	echo json_encode('Désolé une erreur est survenue si celle-ci persiste veuillez la signaler');
 }
 ?>
